pyproject.toml
[project]
name = "invoice-act-tracker"
version = "0.1.0"
description = "Web application for tracking invoices and acts"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "openpyxl>=3.1.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "workalendar>=17.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.2.0",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.1",
]
